requirements.txt
Flask
flask-cors
pymongo
python-dotenv
Werkzeug
groq
transformers
torch
opencv-python
numpy
Pillow
deepface
SpeechRecognition
pydub